Move some code to make it reusable without import cycles.
