+++ api/internal/generators/secret.go
@@ -4,7 +4,6 @@
 package generators
 
 import (
-	"sigs.k8s.io/kustomize/api/filters/filtersutil"
 	"sigs.k8s.io/kustomize/api/ifc"
 	"sigs.k8s.io/kustomize/api/types"
 	"sigs.k8s.io/kustomize/kyaml/yaml"
@@ -51,14 +50,9 @@ func MakeSecret(
 	if err != nil {
 		return nil, err
 	}
-	for _, k := range filtersutil.SortedMapKeys(m) {
-		vrN := makeSecretValueRNode(m[k])
-		if _, err = rn.Pipe(
-			yaml.LookupCreate(yaml.MappingNode, yaml.DataField),
-			yaml.SetField(k, vrN)); err != nil {
-			return nil, err
-		}
+	if err = rn.LoadMapIntoSecretData(m); err != nil {
+		return nil, err
 	}
 	copyLabelsAndAnnotations(rn, args.Options)
-	return rn, err
+	return rn, nil
 }
